Decouple NewVolumeCustomizer and ExtraHddBodyEffector
 Volume specific methods used by NewVolumeCustomizer ExtraHddBodyEffector
 are consumed only from VolumeManager instance.
