+++ blockstore/src/main/java/brooklyn/location/blockstore/api/VolumeManager.java
@@ -13,6 +13,13 @@ import brooklyn.location.blockstore.FilesystemOptions;
 public interface VolumeManager {
 
     /**
+     * Creates a new volume in the availability zone of the machine.
+     * @param machine
+     * @param volumeOptions
+     */
+    MountedBlockDevice createAndAttachDisk(JcloudsMachineLocation machine, VolumeOptions volumeOptions);
+
+    /**
      * Creates a new volume in the given availability zone.
      *
      * @param location Location where the volume should be created
